[BE] refactor : 전체 게시글 조회 수정, 토큰 검증 로직 추가
