+++ server/src/main/java/com/mainproject/auth/AuthController.java
@@ -1,9 +1,12 @@
 package com.mainproject.auth;
 
+import com.mainproject.auth.jwt.JwtTokenizer;
 import com.mainproject.member.entity.Member;
 import com.mainproject.member.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -50,6 +53,14 @@ public class AuthController {
         response.addHeader("Authorization", "Bearer " + newAccessToken);
 
         return ResponseEntity.ok().build();
+    }
+
+    @GetMapping("/access")
+    public ResponseEntity verifiedAccessToken(@AuthenticationPrincipal String email) {
+
+        memberService.findMemberByEmail(email);
+
+        return ResponseEntity.ok().body("Access token validation successful!");
     }
 }
