+++ server/src/main/java/com/mainproject/member/service/MemberService.java
@@ -1,6 +1,6 @@
 package com.mainproject.member.service;
 
-import com.mainproject.auth.CustomAuthorityUtils;
+import com.mainproject.auth.utils.CustomAuthorityUtils;
 import com.mainproject.comment.entity.Comment;
 import com.mainproject.global.exception.BusinessLogicException;
 import com.mainproject.global.exception.ExceptionCode;
