+++ server/src/main/java/com/mainproject/post/repository/PostRepository.java
@@ -9,13 +9,13 @@ import java.util.List;
 
 public interface PostRepository extends JpaRepository<Post,Long> {
 
-    Page<Post> findByTitleContainingAndPostTypeContainingAndMedicalTag_titleContainingAndRegion_nameContainingAndPostStatusNotIn(String keyword, String postType, String name, String title, List<Post.PostStatus> status, Pageable pageable);
+    Page<Post> findByTitleContainingAndPostTypeAndMedicalTag_titleContainingAndRegion_nameContainingAndPostStatusNotIn(String keyword, String postType, String name, String title, List<Post.PostStatus> status, Pageable pageable);
 
-    Page<Post> findByContentContainingAndPostTypeContainingAndMedicalTag_titleContainingAndRegion_nameContainingAndPostStatusNotIn(String keyword, String postType, String name, String title, List<Post.PostStatus> status, Pageable pageable);
+    Page<Post> findByContentContainingAndPostTypeAndMedicalTag_titleContainingAndRegion_nameContainingAndPostStatusNotIn(String keyword, String postType, String name, String title, List<Post.PostStatus> status, Pageable pageable);
 
-    Page<Post> findByMember_displayNameAndPostTypeContainingAndMedicalTag_titleContainingAndRegion_nameContainingAndPostStatusNotIn(String keyword, String postType, String name, String title, List<Post.PostStatus> status, Pageable pageable);
+    Page<Post> findByMember_displayNameContainingAndPostTypeAndMedicalTag_titleContainingAndRegion_nameContainingAndPostStatusNotIn(String keyword, String postType, String name, String title, List<Post.PostStatus> status, Pageable pageable);
 
-    Page<Post> findByTitleContainingAndContentContainingAndPostTypeContainingAndMedicalTag_titleContainingAndRegion_nameContainingAndPostStatusNotIn(String titleKeyword, String contentKeyword, String postType, String name, String title, List<Post.PostStatus> status, Pageable pageable);
+    Page<Post> findByTitleContainingAndContentContainingAndPostTypeAndMedicalTag_titleContainingAndRegion_nameContainingAndPostStatusNotIn(String titleKeyword, String contentKeyword, String postType, String name, String title, List<Post.PostStatus> status, Pageable pageable);
 
     List<Post> findByPostStatus(Post.PostStatus postStatus);
